Adjust RoundGenerator to consider Math.Round methods
