+++ src/NHibernate/Linq/Functions/MathGenerator.cs
@@ -45,16 +45,14 @@ namespace NHibernate.Linq.Functions
 				ReflectHelper.GetMethodDefinition(() => Math.Sign(default(short))),
 				ReflectHelper.GetMethodDefinition(() => Math.Sign(default(sbyte))),
 
-				ReflectHelper.GetMethodDefinition(() => Math.Round(default(decimal))),
-				ReflectHelper.GetMethodDefinition(() => Math.Round(default(decimal), default(int))),
-				ReflectHelper.GetMethodDefinition(() => Math.Round(default(double))),
-				ReflectHelper.GetMethodDefinition(() => Math.Round(default(double), default(int))),
 				ReflectHelper.GetMethodDefinition(() => Math.Floor(default(decimal))),
 				ReflectHelper.GetMethodDefinition(() => Math.Floor(default(double))),
 				ReflectHelper.GetMethodDefinition(() => decimal.Floor(default(decimal))),
+				
 				ReflectHelper.GetMethodDefinition(() => Math.Ceiling(default(decimal))),
 				ReflectHelper.GetMethodDefinition(() => Math.Ceiling(default(double))),
 				ReflectHelper.GetMethodDefinition(() => decimal.Ceiling(default(decimal))),
+				
 				ReflectHelper.GetMethodDefinition(() => Math.Truncate(default(decimal))),
 				ReflectHelper.GetMethodDefinition(() => Math.Truncate(default(double))),
